# reuse the Newtonsoft.Json serializer