Rather than call `JsonConvert` on each instance, which creates a new `JsonSerializer` instance internally - let's just re-use the one created when the `NewtonsoftJsonSerializer` is instantiated.

Noticed this while working on DU serialization.

## Before
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1110 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.302
  [Host]     : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT
  DefaultJob : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT


```
|                            Method |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|---------------------------------- |---------:|---------:|---------:|-------:|----------:|
| Json_serializer_message_roundtrip | 27.83 μs | 0.167 μs | 0.140 μs | 3.2000 |     13 KB |

## After

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1110 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.302
  [Host]     : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT
  DefaultJob : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT


```
|                            Method |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|---------------------------------- |---------:|---------:|---------:|-------:|----------:|
| Json_serializer_message_roundtrip | 26.36 μs | 0.125 μs | 0.104 μs | 3.0000 |     13 KB |


Looks like it helped a small amount (looking at the Gen 0 numbers)