@to11mtm 2021-08-12
There are thread hazards around re-using a JsonSerializer with certain configurations, IIRC specifically around TypeName resolution and/or Preserving object references. When I did #4929 I found any attempts to re-use the serializer would result in a lot of red on my unit tests. [There's a PR to fix in Json.NET](https://github.com/JamesNK/Newtonsoft.Json/pull/2170) but it has been in limbo for some time..